Address PR review feedback (#8062)
- preserve provider kill switches on implicit Codex auth resolution
- disable the unsupported cache-key extension for Groq
- make the Codex model probe fixture realistic and deterministic
- extract oversized provider and resolution test modules
